# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Autoquímicos - Sistema de Gestión de Salas
  description: |
    API REST para el sistema de gestión de salas de conferencias de Autoquímicos S.A. de C.V.
    
    ## Características
    - Gestión de usuarios con roles (Administrador, Aprobador, Organizador, Recepción)
    - Reservas de salas con flujo de aprobación
    - Sistema de coffee break automático
    - Check-in y control de asistencia
    - Auditoría completa de acciones
    - Reportes y estadísticas
    
    ## Autenticación
    La API utiliza JWT (JSON Web Tokens) para autenticación. Incluye el token en el header:
    ```
    Authorization: Bearer <token>
    ```
    
  version: 1.0.0
  contact:
    name: Autoquímicos S.A. de C.V.
    email: admin@autoquimicos.com
  license:
    name: Propietario
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://api.autoquimicos.com/api
    description: Servidor de producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Room:
      type: object
      properties:
        id_sala:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Sala Ejecutiva"
        capacidad:
          type: integer
          example: 12
        descripcion:
          type: string
          example: "Sala principal para reuniones ejecutivas"
        activa:
          type: boolean
          example: true
...
paths:
...
  /rooms:
    get:
      tags:
        - Salas
      summary: Listar salas
      description: Obtiene todas las salas activas
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de salas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 2
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
  /rooms/available/search:
    get:
      tags:
        - Salas
      summary: Buscar salas disponibles
      description: Busca salas disponibles para un horario específico
      security:
        - BearerAuth: []
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-10"
        - name: hora_inicio
          in: query
          required: true
          schema:
            type: string
            format: time
          example: "09:00:00"
        - name: hora_fin
          in: query
          required: true
          schema:
            type: string
            format: time
          example: "11:00:00"
        - name: minCapacity
          in: query
          schema:
            type: integer
          example: 8
      responses:
        '200':
          description: Salas disponibles encontradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 1
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
...